% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confInv}
\alias{confInv}
\title{compute score functions for proportion of variation}
\usage{
confInv(range_h, y, X, lambda, tolerance = 1e-04, confLevel = 0.95)
}
\arguments{
\item{y}{A n\%*1 vector of observed responses}

\item{X}{A n\%*p predictors matrix, n > p}

\item{lambda}{A n\%*1 vector represent values in the diagonal matrix Lambda}

\item{confLevel}{description}
}
\value{
A single value showing the score test statistics
}
\description{
{

}
}
\details{
{
}
}
