% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confReg}
\alias{confReg}
\title{compute score functions for proportion of variation}
\usage{
confReg(range_h, range_p, y, X, lambda, grid = 200L)
}
\arguments{
\item{range_p}{description}

\item{y}{A n\%*1 vector of observed responses}

\item{X}{A n\%*p predictors matrix, n > p}

\item{lambda}{A n\%*1 vector represent values in the diagonal matrix Lambda}

\item{grid}{description}
}
\value{
A single value showing the score test statistics
}
\description{
{

}
}
\details{
{
}
}
