\name{lmmvar-package}
\alias{lmmvar-package}
\alias{lmmvar}
\docType{package}
\title{
\packageTitle{lmmvar}
}
\description{
\packageDescription{lmmvar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lmmvar}
\packageIndices{lmmvar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lmmvar}

Maintainer: \packageMaintainer{lmmvar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
