% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lmmvar}
\alias{lmmvar}
\alias{<lmmvar>-package}
\title{lmmvar}
\description{
compute score functions for 1d and 2d, give confidence interval for proportion of variation, and a confidence grid matrix for proportion of variation and total variation
}
\author{
yiqiao zhang <yqzhang5972 at outlook.com>
}
