% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRTCI.R
\name{RLRTCI}
\alias{RLRTCI}
\title{Confidence interval for variance ratio using restricted likelihood ratio test method.}
\usage{
RLRTCI(SimDists, ciseq, Xnew, ynew, eigens, alpha = 0.05)
}
\arguments{
\item{SimDists}{A vector giving the simulated distribution of RLRT statistic.}

\item{ciseq}{A vector values that need to be checked for \eqn{ sigma^2_g / sigma^2_e}.}

\item{Xnew}{An n-by-p matrix of predictors, n > p. Value needs to be transformed by the same way.}

\item{ynew}{A vector of length n of observed responses. Value needs to be transformed so that covariance matrix of ynew is diagonal.}

\item{eigens}{A vector giving eigen values of covariance matrix of random component in the model.}

\item{alpha}{A scalar in \eqn{(0,1)} indicating confidence level, default is 0.05.}
}
\value{
A vector of length 2 with endpoints of the confidence interval.
}
\description{
This function gives a confidence interval based on Simulated distribution and observed RLRT statistics, see reference for method's detail.
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 \Lambda + \sigma_e^2 I_n),}, where \eqn{\Lambda} is a diagonal matrix with non-negative diagonal elements
supplied in the argument vector \code{lambda}. The parameter of interest is \eqn{ sigma^2_g / sigma^2_e}.
}
\references{
Crainiceanu, C. and Ruppert, D. (2004) Likelihood ratio tests in
linear mixed models with one variance component, \emph{Journal of the Royal
Statistical Society: Series B},\bold{66},165--185.
}
