% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRTCI.R
\name{RLRTSim}
\alias{RLRTSim}
\title{Simulated distribution of the restricted likelihood ratio test statistic of variance ratio.}
\usage{
RLRTSim(
  X,
  ZSigmasqrt,
  tau0 = 0,
  nsim = 10000,
  seed = NA,
  grid = 200,
  tolerate = 1e-05
)
}
\arguments{
\item{X}{An n-by-p matrix of predictors, n > p.}

\item{ZSigmasqrt}{An n-by-m matrix with value equal to square root of covariance matrix of random component.}

\item{tau0}{A scalar in \eqn{(0, \infty)} represent the value we want to test, dafault is 0.}

\item{nsim}{A positive integer indicating how many simulations you want.}

\item{seed}{A scalar to allow setting seed, default is NA.}

\item{grid}{A positive integer indicating how many grid you want.}

\item{tolerate}{A positive scalar indicating how close to boundary \eqn{\tau} needs to be.}
}
\value{
A vector of length nsim.
}
\description{
Form a simulated distribution of RLRT statistic when testing \eqn{\sigma_g^2/\sigma_e^2 = \tau_0}, see reference for method's detail.
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 K + \sigma_e^2 I_n),}, where \eqn{K=ZZ^T} is the covariance matrix of random component.
The parameter of interest is \eqn{ sigma^2_g / sigma^2_e}.
}
\references{
Crainiceanu, C. and Ruppert, D. (2004) Likelihood ratio tests in
linear mixed models with one variance component, \emph{Journal of the Royal
Statistical Society: Series B},\bold{66},165--185.
}
