% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLRTCI.R
\name{simulate_RLRT}
\alias{simulate_RLRT}
\title{Simulate the distribution of the restricted likelihood ratio test statistic for a variance ratio.}
\usage{
simulate_RLRT(
  X,
  Ksqrt,
  tau0 = 0,
  nsim = 10000,
  seed = NA,
  grid = 200,
  tol = 1e-05
)
}
\arguments{
\item{X}{An n-by-p matrix of predictors, n > p.}

\item{Ksqrt}{An n-by-m matrix that is a square-root of the variance component covariance matrix (see details).}

\item{tau0}{A scalar null hyothesis in \eqn{[0, \infty)}, default is 0.}

\item{nsim}{A positive integer indicating the number of simulations.}

\item{seed}{A scalar to allow setting seed, default is NA.}

\item{grid}{A positive integer with the number of grid points at which to evaluate when inverting the test-statistic to get a confidence region.}

\item{tol}{A positive scalar such that \eqn{h^2 = \tau / (1 + \tau) < 1 - \mathtt{tol}}.}
}
\value{
A vector of length \code{nsim} of simulated values of the test-statistic.
}
\description{
Simulate the distribution of RLRT statistic in a variance component model under the null hypothesis that that \eqn{\tau = \sigma_g^2/\sigma_e^2 = \mathtt{tau0}} (see details and references).
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 K + \sigma_e^2 I_n),} where \eqn{K} is
the variance component covariance matrix. The matrix \code{Ksqrt} should be such
that \code{tcrossprod(Ksqrt)} is equal to \eqn{K}.

The parameter of interest is \eqn{\tau = \sigma^2_g / \sigma^2_e}.

An implementation of this test-statistic for the case where \code{tau0 = 0}
is also available in the package \code{RLRsim}.
}
\references{
Crainiceanu, C. and Ruppert, D. (2004) Likelihood ratio tests in
linear mixed models with one variance component, \emph{Journal of the Royal
Statistical Society: Series B},\bold{66},165--185.
}
