% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confReg}
\alias{confReg}
\title{Joint confidence region for proportion of variation, or heritability, and total variation}
\usage{
confReg(
  y,
  X,
  lambda,
  range_h = as.numeric(c(0, 1)),
  range_p = as.numeric(c(0, 1)),
  grid = 200L
)
}
\arguments{
\item{y}{A vector of length n of observed responses with diagonal covariance matrix.}

\item{X}{An n-by-p matrix of predictors, n > p.}

\item{lambda}{A vector of length n with (non-negative) variances, which are the
eigenvalues of the variance component covariance matrix (see details).}

\item{range_h}{A vector of length 2 with the boundaries for \eqn{h^2}.
The endpoints must be in [0,1).}

\item{range_p}{A vector of length 2 giving the boundaries for \eqn{\sigma^2_p}.
The endpoints must be positive.}

\item{grid}{The number of grid points in each interval, meaning the total number
of points in the grid is \code{grid^2}.}
}
\value{
A \code{grid}-by-\code{grid} matrix with the test-statistic evaluated at the corresponding
grid points. Rows index \code{h2}, columns index \code{s2p}.
}
\description{
Calculate a confidence region by numerically inverting the test-statistic
implemented in the function \code{varRatioTest2d}. Numerical inversion is done
by evaluating the test-statistic on a grid (see details).
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 \Lambda + \sigma_e^2 I_n),}
where \eqn{\Lambda} is a diagonal matrix with non-negative diagonal elements
supplied in the argument vector \code{lambda}. See the documentation for
\code{varRatioTest1d} for how this model often results from transforming more
common ones using an eigendecomposition.

The parameters of interest are \eqn{h^2 = \sigma^2_g / \sigma^2_p}
and \eqn{\sigma^2_p = \sigma^2_g + \sigma^2_e}.

The function creates a set of feasible values for \eqn{h^2} by taking \code{grid}
evenly spaced points in the interval defined by \code{range_h}. It creates
a set of feasible values for \eqn{\sigma^2_p} by doing the same thing
using the interval defined by \code{range_p}. The Cartesian product of these
two sets is the grid on which the test-statistic is evaluated. A point on
this grid is in the confidence region if the value of the test-statistic
is less than the appropriate quantile of the chi-square distribution with
2 degrees of freedom.
}
