% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmvar-package.R
\docType{package}
\name{lmmvar}
\alias{lmmvar}
\alias{lmmvar-package}
\title{lmmvar}
\description{
compute score functions for 1d and 2d, give confidence interval for proportion of variation, and a confidence grid matrix for proportion of variation and total variation
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yqzhang5972/lmmvar}
  \item Report bugs at \url{https://github.com/yqzhang5972/lmmvar/issues}
}

}
\author{
yiqiao zhang \if{html}{\out{<yqzhang5972 at outlook.com>}}
}
