% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRatioTest1d}
\alias{varRatioTest1d}
\title{1d score test statistics for proportion of variation}
\usage{
varRatioTest1d(h2, y, X, lambda, sqRoot = FALSE)
}
\arguments{
\item{h2}{an numeric indicates the desired value of proportion of variation that need to be tested. Needs to be within [0,1).}

\item{y}{A n\eqn{\times}1 vector of observed responses.}

\item{X}{A n\eqn{\times}p predictors matrix, n > p.}

\item{lambda}{A n\eqn{\times}1 vector represent values in the diagonal matrix Lambda. Values need to be non-negative.}

\item{sqRoot}{A boolean indicates whether to return score test statistic or +/- square root of it (\eqn{I^{-1/2}U}). Default is FALSE.}
}
\value{
A single value showing the score test statistics at h2.
}
\description{
{
compute the score test statistic for a proportion of variation of indepedent random component versus total variation in a linear mixed model,
assuming the covariance matrix corresponding to the random component is diagonal (see details).
}
}
\details{
{ Assuming the linear mixed model follows \eqn{y \sim N(X\beta, \sigma_g\Lambda + \sigma_e I_n)}.
The proportion of variation of indepedent random component, \eqn{h^2}, is \eqn{\sigma_g / (\sigma_g+\sigma_e)},
the total variation \eqn{\sigma_p = \sigma_g+\sigma_e}, then y can also be seen to follow \eqn{N(X\beta, \sigma_p(h^2\Lambda + (1-h^2)I_n))}.
\eqn{\Lambda} is a diagonal matrix which can be achieved by applying eigen decomposition to your non-diagonal SPD \eqn{\Lambda} as well as X and y.
By the nature of the model, the support set of h2 has to be in [0,1). The test statistic follows \eqn{\Chi_1}.
}
}
