% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRatioTest1d}
\alias{varRatioTest1d}
\title{1d score test statistics for proportion of variation}
\usage{
varRatioTest1d(h2, y, X, lambda)
}
\arguments{
\item{h2}{an numeric indicates the desired value of proportion of variation that need to be tested. Needs to be within [0,1).}

\item{y}{A n\%*1 vector of observed responses.}

\item{X}{A n\%*p predictors matrix, n > p.}

\item{lambda}{A n\%*1 vector represent values in the diagonal matrix Lambda. Values need to be non-negative.}
}
\value{
A single value showing the score test statistics at h2.
}
\description{
{
compute the score test statistic for a proportion of variation of indepedent random component versus total variation in a linear mixed model,
assuming the covariance matrix corresponding to the random component is diagonal (see details).
}
}
\details{
{ Assuming the linear mixed model follows y ~ N(X\% beta, \% sigma_g \% Lambda + \% sigma_e I).
The proportion of variation of indepedent random component, h2, is \% sigma_g / (\% sigma_g+\% sigma_e),
the total variation \% sigma_p = \% sigma_g+\% sigma_e, then y can also be seen to follow N(X\% beta, \% sigma_p(h2\% Lambda + (1-h2)I)).
\% Lambda is a diagonal matrix which can be achieved by applying eigen decomposition to your non-diagonal SPD \% Lambda as well as X and y.
By the nature of the model, the support set of h2 has to be in [0,1). The test statistic follows \% Chi(1).
}
}
