% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRatioTest1d}
\alias{varRatioTest1d}
\title{Restricted score test-statistic for a proportion of variation, or heritability}
\usage{
varRatioTest1d(h2, y, X, lambda, sqRoot = FALSE)
}
\arguments{
\item{h2}{The null hypothesis value, which needs to be in [0,1).}

\item{y}{A vector of length n of observed responses with diagonal covariance matrix.}

\item{X}{An n-by-p matrix of predictors, n > p.}

\item{lambda}{A vector of length n with (non-negative) variances, which are the
eigenvalues of the variance component covariance matrix (see details).}

\item{sqRoot}{If \code{true}, return statistic for signed square root statistic. Defaults to \code{false}.}
}
\value{
The test-statistic evaluated at \code{h2}.
}
\description{
Compute a restricted score test-statistic for the proportion of variation due to the
variance component in a model with one variance component and an error term.
The function assumes the covariance matrix corresponding to the variance component is
diagonal which in practice usually means the actual covariance matrix
has been eigendecomposed and the transformed data are supplied to this
function (see details). This proportion is known as heritability in some
applications, and therefore denoted \code{h2} in the code.
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 \Lambda + \sigma_e^2 I_n),}
where \eqn{\Lambda} is a diagonal matrix with non-negative diagonal elements
supplied in the argument vector \code{lambda}. The parameter of interest is
\eqn{h^2=\sigma_g^2/(\sigma^2_g + \sigma^2_e)}.

Usually this model results
from transformation: If
\deqn{\tilde{y} \sim N(\tilde{X} \beta, \sigma_g^2 K + \sigma_e^2 I_n),}
where \eqn{K} is a positive semi-definite (covariance) matrix with
eigendecomposition \eqn{U\Lambda U^\top}, then the transformed responses
\eqn{y = U^\top \tilde{y}} and predictors \eqn{X = U^\top \tilde{X}} satisfy
the model the function assumes.

A linear mixed model with one random effect,
\eqn{\tilde{y} = \tilde{X}\beta + ZU + E}, where \eqn{U\sim N(0, \sigma^2_g I_q)}
and \eqn{E \sim N(0, \sigma^2_e I_n)}, is equivalent to the above with
\eqn{K = ZZ^\top}.

The test-statistic is approximately chi-square with one degree of
freedom, even if \code{h2} is small or equal to zero, that is, near or at the
boundary of the parameter set. If `sqRoot = TRUE',
then the test-statistic is approximately standard normal.

If the parameter of interest is instead \eqn{\tau = \sigma^2_g/\sigma^2_e},
note \eqn{h^2 = \tau / (1 + \tau)}, so the function can be evaluated the
null hypothesis value for \eqn{\tau}, say \code{tau}, by calling
\code{varRatioTest1d(h2 = tau / (1 + tau), ...)}.
}
