% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRatioTest2d}
\alias{varRatioTest2d}
\title{compute score functions for proportion of variation}
\usage{
varRatioTest2d(h2, s2p, y, X, lambda)
}
\arguments{
\item{h2}{an numeric indicates the desired value of h2 that need to be tested}

\item{s2p}{description}

\item{y}{A n\%*1 vector of observed responses}

\item{X}{A n\%*p predictors matrix, n > p}

\item{lambda}{A n\%*1 vector represent values in the diagonal matrix Lambda}
}
\value{
A single value showing the score test statistics
}
\description{
{

}
}
\details{
{
}
}
