% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRatioTest2d}
\alias{varRatioTest2d}
\title{Joint restricted score test for a proportion of variation, or heritability, and total variation}
\usage{
varRatioTest2d(h2, s2p, y, X, lambda)
}
\arguments{
\item{h2}{The null hypothesis value of \eqn{h^2}, which needs to be in [0,1).}

\item{s2p}{The null hypothesis value of \eqn{\sigma^2_p = \sigma^2_g + \sigma^2_e}.}

\item{y}{A vector of length n of observed responses with diagonal covariance matrix.}

\item{X}{An n-by-p matrix of predictors, n > p.}

\item{lambda}{A vector of length n with (non-negative) variances, which are the
eigenvalues of the variance component covariance matrix (see details).}
}
\value{
The test-statistic evaluated at \code{h2} and \code{s2p}.
}
\description{
Compute a joint restricted score test-statistic for the total variance and the proportion
of variation due to the variance component in a model with one variance component and an error term.
See details below, and the documentation of varRatioTest1d.
}
\details{
The function assumes the model
\deqn{y \sim N(X \beta, \sigma_g^2 \Lambda + \sigma_e^2 I_n),}
where \eqn{\Lambda} is a diagonal matrix with non-negative diagonal elements
supplied in the argument vector \code{lambda}. See the documentation for
\code{varRatioTest1d} for how this model often results from transforming more
common ones using an eigendecomposition.

The parameters in the function are \eqn{h^2=\sigma_g^2/\sigma^2_p} and
\eqn{\sigma^2_p = \sigma^2_g + \sigma^2_e}.

The test-statistic is approximately chi-square with two degrees of
freedom, even if \code{h2} and \code{s2p} are small.
}
